+++ packages/ui-kit-js/src/components/navigation/link/link.tsx
@@ -1,9 +1,10 @@
-import { JSX, Component, Prop, h, Element, Listen } from '@stencil/core';
+import { Component, Element, h, JSX, Prop } from '@stencil/core';
 import cx from 'classnames';
 import { BreakpointCustomizable, mapBreakpointPropToPrefixedClasses, prefix } from '../../../utils';
 import { IconName } from '../../icon/icon/icon-name';
-import { improveFocusHandlingForCustomElement, preventNativeTabIndex } from '../../../utils/focusHandling';
-import { Theme, LinkTarget } from '../../../types';
+import { improveFocusHandlingForCustomElement } from '../../../utils/focusHandling';
+import { LinkTarget, Theme } from '../../../types';
+import { improveLinkHandlingForCustomElement } from '../../../utils/linkHandling';
 
 @Component({
   tag: 'p-link',
@@ -12,19 +13,6 @@ import { Theme, LinkTarget } from '../../../types';
 })
 export class Link {
   @Element() public element!: HTMLElement;
-
-  /**
-   * Check native tabindex to ensure that it doesn't get set on the host element
-   * @internal
-   */
-  @Prop({
-    mutable: true,
-    attribute: 'tabindex'
-  })
-  public nativeTabindex?: number = -1;
-
-  /** To remove the element from tab order. */
-  @Prop() public tabbable?: boolean = true;
 
   /** The style variant of the link. */
   @Prop() public variant?: 'primary' | 'secondary' | 'tertiary' = 'secondary';
@@ -55,11 +43,10 @@ export class Link {
 
   public componentDidLoad() {
     improveFocusHandlingForCustomElement(this.element);
+    improveLinkHandlingForCustomElement(this.element);
   }
 
   public render(): JSX.Element {
-    preventNativeTabIndex(this);
-
     const TagType = this.href === undefined ? 'span' : 'a';
 
     const linkClasses = cx(
@@ -74,9 +61,12 @@ export class Link {
     return (
       <TagType
         class={linkClasses}
-        href={this.href}
-        {...(TagType === 'a' ? { href: this.href, target: `${this.target}`, download: this.download, rel: this.rel } : null)}
-        {...(TagType === 'a' && this.tabbable ? { tabindex: 0 } : { tabindex: -1 })}
+        {...(TagType === 'a' ? {
+          href: this.href,
+          target: `${this.target}`,
+          download: this.download,
+          rel: this.rel
+        } : null)}
       >
         <p-icon
           class={iconClasses}
@@ -89,20 +79,5 @@ export class Link {
         </p-text>
       </TagType>
     );
-  }
-
-
-  /**
-   * IE11 workaround to fix the event target
-   * of click events (which normally shadow dom
-   * takes care of)
-   */
-  @Listen('click', { capture: true })
-  public fixEventTarget(event: MouseEvent): void {
-    if (event.target !== this.element) {
-      event.stopPropagation();
-      event.preventDefault();
-      this.element.click();
-    }
   }
 }